Resize image
Original image was 12.76MB - new image 82.24KB (155x smaller)
Webpage optimization is to use minimal resources - serving large images is of no use in webpages where >90% is on a screen resolution of FullHD(1920x1080) or a mobile phone. Where such an image will not show it's full potential.

The original image will not be accessible from this commit forward, but you can go through the history to find the original image back.